requirements.txt
certifi==2017.4.17
chardet==3.0.4
construct==2.5.3
Django==1.11.3
django-admin-tools==0.8.1
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-js-asset==0.1.1
django-meta==1.3.2
django-mptt==0.8.7
django-mptt-admin==0.4.6
future==0.16.0
haystack==0.42
idna==2.5
lxml==3.8.0
pefile==2016.3.28
psycopg2==2.7.2
python-ptrace==0.9.2
pytz==2017.2
requests==2.18.1
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.3
urllib3==1.21.1
wget==3.2
